+++ test/presentation/page/repo/component/repo_list_view_test.dart
@@ -9,8 +9,8 @@ import 'package:github_search/infrastructure/github/http_client.dart';
 import 'package:github_search/presentation/component/error_view.dart';
 import 'package:github_search/presentation/component/list_loader.dart';
 import 'package:github_search/presentation/page/repo/component/repo_list_view.dart';
+import 'package:github_search/presentation/router.dart';
 import 'package:github_search/util/localization/strings.g.dart';
-import 'package:github_search/util/routing/router.dart';
 import 'package:go_router/go_router.dart';
 import 'package:number_display/number_display.dart';
 import 'package:visibility_detector/visibility_detector.dart';
